job_automation/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.1.0
numpy>=1.26.3
pydantic>=2.3.0
tls-client>=1.0.1
markdownify>=1.1.0
regex>=2024.4.28
openai>=1.0.0
python-dotenv>=1.0.0
